jarvis/requirements.txt
# JARVIS Desktop Assistant - Dependencies

# Core
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0

# AI
anthropic>=0.49.0

# Voice
edge-tts>=6.1.0
elevenlabs>=1.0.0

# Google (Gmail + Calendar)
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0

# Slack
slack-sdk>=3.27.0

# Notion
notion-client>=2.2.0

# HubSpot
hubspot-api-client>=9.0.0

# Utilities
httpx>=0.27.0
pydantic>=2.6.0
python-dateutil>=2.9.0
